+++ examples/cylinder/mesh/mesh.jl
@@ -4,9 +4,9 @@
 using Gmsh: gmsh
 
 """
-    generate_cylindrical_cavity_mesh(;
+    generate_cylindrical_mesh(;
         filename::AbstractString,
-        refinement::Integer  = 0,
+        refinement::Integer  = 1,
         order::Integer       = 1,
         mesh_type::Integer   = 0,
         radius::Real         = 2.74,
@@ -16,7 +16,7 @@ using Gmsh: gmsh
         gui::Bool            = false
     )
 
-Generate a mesh for the cylindrical cavity resonator example using Gmsh
+Generate a mesh for the cylinder example using Gmsh
 
 # Arguments
 
@@ -24,15 +24,15 @@ Generate a mesh for the cylindrical cavity resonator example using Gmsh
   - refinement - measure of how many elements to include, 0 is least
   - order - the polynomial order of the approximation, minimum 1
   - mesh_type - 0 = tetrahedral mesh, 1 = prism mesh, 2 = hexahedral mesh
-  - radius - the radius of the cavity resonator
-  - aspect_ratio - the ratio of the diameter (not radius!) of the cavity to the height
+  - radius - the radius of the cylinder
+  - aspect_ratio - the ratio of the diameter (not radius!) of the cylinder to the height
   - symmetry_plane - whether to cut the cylinder in half and use a symmetry plane in the model
   - verbose - flag to dictate the level of print to REPL, passed to Gmsh
   - gui - whether to launch the Gmsh GUI on mesh generation
 """
-function generate_cylindrical_cavity_mesh(;
+function generate_cylindrical_mesh(;
     filename::AbstractString,
-    refinement::Integer  = 0,
+    refinement::Integer  = 1,
     order::Integer       = 1,
     mesh_type::Integer   = 0,
     radius::Real         = 2.74,
@@ -50,11 +50,11 @@ function generate_cylindrical_cavity_mesh(;
     gmsh.option.setNumber("General.Verbosity", verbose)
 
     # Add model
-    if "cavity" in gmsh.model.list()
-        gmsh.model.setCurrent("cavity")
+    if "cylinder" in gmsh.model.list()
+        gmsh.model.setCurrent("cylinder")
         gmsh.model.remove()
     end
-    gmsh.model.add("cavity")
+    gmsh.model.add("cylinder")
 
     # Geometry parameters (in cm)
     height = aspect_ratio * 2 * radius  # Cylinder height
@@ -106,7 +106,31 @@ function generate_cylindrical_cavity_mesh(;
     # Add physical groups
     cylinder_group = gmsh.model.addPhysicalGroup(3, cylinder, -1, "cylinder")
 
-    symmetry =
+    @show top =
+        last.(
+            gmsh.model.getEntitiesInBoundingBox(
+                -1.1 * radius,
+                -1.1 * radius,
+                0.5 * height,
+                1.1 * radius,
+                1.1 * radius,
+                1.1 * height,
+                2
+            )
+        )
+    @show bottom =
+        last.(
+            gmsh.model.getEntitiesInBoundingBox(
+                -1.1 * radius,
+                -1.1 * radius,
+                -0.1 * height,
+                1.1 * radius,
+                1.1 * radius,
+                0.5 * height,
+                2
+            )
+        )
+    @show symmetry =
         last.(
             gmsh.model.getEntitiesInBoundingBox(
                 -1.1 * radius,
@@ -118,22 +142,24 @@ function generate_cylindrical_cavity_mesh(;
                 2
             )
         )
-    boundaries = []
+    exterior = []
     for domain in cylinder
         _, domain_boundaries = gmsh.model.getAdjacencies(3, domain)
         for boundary in domain_boundaries
-            if boundary in symmetry
+            if boundary in top || boundary in bottom || boundary in symmetry
                 continue
             end
-            idx = first(indexin([boundary], boundaries))
+            idx = first(indexin([boundary], exterior))
             if isnothing(idx)
-                push!(boundaries, boundary)
+                push!(exterior, boundary)
             else
-                deleteat!(boundaries, idx)
+                deleteat!(exterior, idx)
             end
         end
     end
-    boundary_group = gmsh.model.addPhysicalGroup(2, boundaries, -1, "boundaries")
+    top_group = gmsh.model.addPhysicalGroup(2, top, -1, "top")
+    bottom_group = gmsh.model.addPhysicalGroup(2, bottom, -1, "bottom")
+    exterior_group = gmsh.model.addPhysicalGroup(2, exterior, -1, "exterior")
     symmetry_group = gmsh.model.addPhysicalGroup(2, symmetry, -1, "symmetry")
 
     # Generate mesh
@@ -180,7 +206,9 @@ function generate_cylindrical_cavity_mesh(;
     if verbose > 0
         println("\nFinished generating mesh. Physical group tags:")
         println("Cylinder: ", cylinder_group)
-        println("Boundaries: ", boundary_group)
+        println("Exterior boundaries: ", exterior_group)
+        println("Top boundaries: ", top_group)
+        println("Bottom boundaries: ", bottom_group)
         if length(symmetry) > 0
             println("Symmetry boundaries: ", symmetry_group)
         end