+++ palace/fem/libceed/restriction.cpp
@@ -12,16 +12,105 @@ namespace palace::ceed
 namespace
 {
 
+const mfem::FiniteElement *GetTraceElement(const mfem::FiniteElementSpace &fespace,
+                                           const std::vector<int> &indices)
+{
+  int elem_id, face_info;
+  fespace.GetMesh()->GetBdrElementAdjacentElement(indices[0], elem_id, face_info);
+  mfem::Geometry::Type face_geom = fespace.GetMesh()->GetBdrElementGeometry(indices[0]);
+  return fespace.GetTraceElement(elem_id, face_geom);
+};
+
+mfem::Array<int> GetFaceDofsFromAdjacentElement(const mfem::FiniteElementSpace &fespace,
+                                                mfem::DofTransformation &dof_trans,
+                                                const int P, const int e)
+{
+  // Get coordinates of face dofs
+  int elem_id, face_info;
+  fespace.GetMesh()->GetBdrElementAdjacentElement(e, elem_id, face_info);
+  mfem::Geometry::Type face_geom = fespace.GetMesh()->GetBdrElementGeometry(e);
+  face_info = fespace.GetMesh()->EncodeFaceInfo(
+      fespace.GetMesh()->DecodeFaceInfoLocalIndex(face_info),
+      mfem::Geometry::GetInverseOrientation(
+          face_geom, fespace.GetMesh()->DecodeFaceInfoOrientation(face_info)));
+  mfem::IntegrationPointTransformation Loc1;
+  fespace.GetMesh()->GetLocalFaceTransformation(fespace.GetMesh()->GetBdrElementType(e),
+                                                fespace.GetMesh()->GetElementType(elem_id),
+                                                Loc1.Transf, face_info);
+  const mfem::FiniteElement *face_el = fespace.GetTraceElement(elem_id, face_geom);
+  MFEM_VERIFY(dynamic_cast<const mfem::NodalFiniteElement *>(face_el),
+              "Mesh requires nodal Finite Element.");
+  mfem::IntegrationRule face_ir(face_el->GetDof());
+  Loc1.Transf.ElementNo = elem_id;
+  Loc1.Transf.mesh = fespace.GetMesh();
+  Loc1.Transf.ElementType = mfem::ElementTransformation::ELEMENT;
+  Loc1.Transform(face_el->GetNodes(), face_ir);
+  mfem::DenseMatrix face_pm;
+  fespace.GetMesh()->GetNodes()->GetVectorValues(Loc1.Transf, face_ir, face_pm);
+
+  // Get coordinates of element dofs
+  mfem::DenseMatrix elem_pm;
+  const mfem::FiniteElement *fe_elem = fespace.GetFE(elem_id);
+  mfem::ElementTransformation *T = fespace.GetMesh()->GetElementTransformation(elem_id);
+  T->Transform(fe_elem->GetNodes(), elem_pm);
+
+  // Find the dofs
+  double tol = 1E-5;
+  mfem::Array<int> elem_dofs, dofs(P);
+  fespace.GetElementDofs(elem_id, elem_dofs, dof_trans);
+  for (int l = 0; l < P; l++)
+  {
+    double norm2_f = 0.0;
+    for (int m = 0; m < face_pm.Height(); m++)
+    {
+      norm2_f += face_pm(m, l) * face_pm(m, l);
+    }
+    for (int m = 0; m < elem_pm.Width(); m++)
+    {
+      double norm2_e = 0.0;
+      for (int n = 0; n < elem_pm.Height(); n++)
+      {
+        norm2_e += elem_pm(n, m) * elem_pm(n, m);
+      }
+      double relative_tol = tol * std::max(std::max(norm2_f, norm2_e), 1.0E-6);
+      double diff = 0.0;
+      for (int o = 0; o < elem_pm.Height(); o++)
+      {
+        diff += std::fabs(elem_pm(o, m) - face_pm(o, l));
+      }
+      if (diff <= relative_tol)
+      {
+        dofs[l] = elem_dofs[m];
+      }
+    }
+  }
+  return dofs;
+};
+
 void InitLexicoRestr(const mfem::FiniteElementSpace &fespace,
                      const std::vector<int> &indices, bool use_bdr, Ceed ceed,
                      CeedElemRestriction *restr)
 {
   const std::size_t num_elem = indices.size();
-  const mfem::FiniteElement &fe =
-      use_bdr ? *fespace.GetBE(indices[0]) : *fespace.GetFE(indices[0]);
-  const int P = fe.GetDof();
-  const mfem::TensorBasisElement *tfe = dynamic_cast<const mfem::TensorBasisElement *>(&fe);
+  const mfem::FiniteElement *fe;
+  bool face_flg = false;
+  if (!use_bdr)
+  {
+    fe = fespace.GetFE(indices[0]);
+  }
+  else
+  {
+    fe = fespace.GetBE(indices[0]);
+    if (!fe)
+    {
+      fe = GetTraceElement(fespace, indices);
+      face_flg = true;
+    }
+  }
+  const int P = fe->GetDof();
+  const mfem::TensorBasisElement *tfe = dynamic_cast<const mfem::TensorBasisElement *>(fe);
   const mfem::Array<int> &dof_map = tfe->GetDofMap();
+  const bool dof_map_is_identity = dof_map.Size() == 0;
   const CeedInt comp_stride =
       (fespace.GetVDim() == 1 || fespace.GetOrdering() == mfem::Ordering::byVDIM)
           ? 1
@@ -45,7 +134,14 @@ void InitLexicoRestr(const mfem::FiniteElementSpace &fespace,
       const int e = indices[i];
       if (use_bdr)
       {
-        fespace.GetBdrElementDofs(e, dofs, dof_trans);
+        if (!face_flg)
+        {
+          fespace.GetBdrElementDofs(e, dofs, dof_trans);
+        }
+        else
+        {
+          dofs = GetFaceDofsFromAdjacentElement(fespace, dof_trans, P, e);
+        }
       }
       else
       {
@@ -56,7 +152,7 @@ void InitLexicoRestr(const mfem::FiniteElementSpace &fespace,
                   "restriction.");
       for (int j = 0; j < P; j++)
       {
-        const int sdid = dof_map[j];  // signed
+        const int sdid = dof_map_is_identity ? j : dof_map[j];  // signed
         const int did = (sdid >= 0) ? sdid : -1 - sdid;
         const int sgid = dofs[did];  // signed
         const int gid = (sgid >= 0) ? sgid : -1 - sgid;
@@ -90,9 +186,22 @@ void InitNativeRestr(const mfem::FiniteElementSpace &fespace,
                      Ceed ceed, CeedElemRestriction *restr)
 {
   const std::size_t num_elem = indices.size();
-  const mfem::FiniteElement &fe =
-      use_bdr ? *fespace.GetBE(indices[0]) : *fespace.GetFE(indices[0]);
-  const int P = fe.GetDof();
+  const mfem::FiniteElement *fe;
+  bool face_flg = false;
+  if (!use_bdr)
+  {
+    fe = fespace.GetFE(indices[0]);
+  }
+  else
+  {
+    fe = fespace.GetBE(indices[0]);
+    if (!fe)
+    {
+      fe = GetTraceElement(fespace, indices);
+      face_flg = true;
+    }
+  }
+  const int P = fe->GetDof();
   const CeedInt comp_stride =
       (fespace.GetVDim() == 1 || fespace.GetOrdering() == mfem::Ordering::byVDIM)
           ? 1
@@ -105,7 +214,7 @@ void InitNativeRestr(const mfem::FiniteElementSpace &fespace,
     {
       return false;
     }
-    const auto geom = fe.GetGeomType();
+    const auto geom = fe->GetGeomType();
     const auto *dof_trans = fespace.FEColl()->DofTransformationForGeometry(geom);
     return (dof_trans && !dof_trans->IsIdentity());
   }();
@@ -140,7 +249,14 @@ void InitNativeRestr(const mfem::FiniteElementSpace &fespace,
       const auto e = indices[i];
       if (use_bdr)
       {
-        fespace.GetBdrElementDofs(e, dofs, dof_trans);
+        if (!face_flg)
+        {
+          fespace.GetBdrElementDofs(e, dofs, dof_trans);
+        }
+        else
+        {
+          dofs = GetFaceDofsFromAdjacentElement(fespace, dof_trans, P, e);
+        }
       }
       else
       {
@@ -258,11 +374,22 @@ void InitRestriction(const mfem::FiniteElementSpace &fespace,
               << indices[0] << ", " << use_bdr << ", " << is_interp << ", "
               << is_interp_range << ")\n";
   }
-  const mfem::FiniteElement &fe =
-      use_bdr ? *fespace.GetBE(indices[0]) : *fespace.GetFE(indices[0]);
-  const mfem::TensorBasisElement *tfe = dynamic_cast<const mfem::TensorBasisElement *>(&fe);
-  const bool vector = fe.GetRangeType() == mfem::FiniteElement::VECTOR;
-  const bool lexico = (tfe && tfe->GetDofMap().Size() > 0 && !vector && !is_interp);
+  const mfem::FiniteElement *fe;
+  if (!use_bdr)
+  {
+    fe = fespace.GetFE(indices[0]);
+  }
+  else
+  {
+    fe = fespace.GetBE(indices[0]);
+    if (!fe)
+    {
+      fe = GetTraceElement(fespace, indices);
+    }
+  }
+  const mfem::TensorBasisElement *tfe = dynamic_cast<const mfem::TensorBasisElement *>(fe);
+  const bool vector = fe->GetRangeType() == mfem::FiniteElement::VECTOR;
+  const bool lexico = (tfe && !vector && !is_interp);
   if (lexico)
   {
     // Lexicographic ordering using dof_map.