+++ palace/utils/configfile.hpp
@@ -467,6 +467,23 @@ public:
   void SetUp(json &boundaries);
 };
 
+struct PeriodicData
+{
+public:
+  // Vector defining the direction and distance for this periodic boundary condition.
+  std::array<double, 3> translation = {0.0, 0.0, 0.0};
+  // List of boundary donor attributes for this periodic boundary condition.
+  std::vector<int> donor_attributes = {};
+  // List of boundary receiver attributes for this periodic boundary condition.
+  std::vector<int> receiver_attributes = {};
+};
+
+struct PeriodicBoundaryData : public internal::DataVector<PeriodicData>
+{
+public:
+  void SetUp(json &boundaries);
+};
+
 struct WavePortData
 {
 public:
@@ -609,6 +626,7 @@ public:
   LumpedPortBoundaryData lumpedport = {};
   WavePortBoundaryData waveport = {};
   SurfaceCurrentBoundaryData current = {};
+  PeriodicBoundaryData periodic = {};
   BoundaryPostData postpro = {};
 
   void SetUp(json &config);